artifacts/outputs/manual-harness/home/standart/subprojects/desktop_markdown_sync/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "desktop-markdown-sync"
version = "0.1.0"
description = "Two-way Markdown and KDE Plasma virtual desktop synchronization for Kali Linux."
readme = "readme.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "standart", email = "standart@example.com"},
]

[project.optional-dependencies]
dev = [
  "bandit[toml]>=1.7.9",
  "pymarkdownlnt>=0.9.25",
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.11.0",
  "PyYAML>=6.0.2",
  "yamllint>=1.35.1",
]

[project.scripts]
desktop-markdown-sync = "desktop_markdown_sync.cli:main"

[tool.pytest.ini_options]
addopts = "-q --cov=src/desktop_markdown_sync --cov-report=term-missing --cov-report=xml"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/desktop_markdown_sync/fixture_notes.py" = ["E501"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
